runtime % support for FreeRTOS V11
In V11, FreeRTOS changed ulTotalRunTime from an integer to an array of
integers. On SMP, the array has one entry per CPU, but it's still a
one-element array even without SMP support compiled in.

Try finding the length of the array with sizeof(a)/sizeof(a[0]). If that
succeeds, sum all CPU runtimes to get the total runtime. If it fails,
assume V10 or earlier and fall back to the previous technique.