FreeRTOS: disable stack usage on ports w/ xMPUSettings.ulContext[] (#65)
On many ARM ports, when `configENABLE_MPU` is used, `pxTopOfStack`
doesn't actually point to the top of the stack. It points into a
separate `pTCB->xMPUSettings.ulContext[]` array that is used to store
the task's registers for context switches. As a result, computing "Stack
Used" and "Stack Free" as deltas between `pxTopOfStack` and `pxStack` /
`pxEndOfStack` gives incorrect, meaningless results.