Hide QUEUES and MUX/SEMS tabs if configQUEUE_REGISTRY_SIZE == 0
There is a hint at the bottom of the panel explaining that
configQUEUE_REGISTRY_SIZE can be set to > 0 to get information about
queues and semaphores, so the option to see these tabs is still
discoverable. For users who have opted out of the queue registry, this
removes the visual clutter of tabs they can't use.