Read stackTop from FreeRTOS MPU context block (#65) (#68)
* Read stack pointer from FreeRTOS MPU context block (#65)

On ARM ports, when `configENABLE_MPU` is used, `pxTopOfStack` doesn't
actually point to the top of the stack. It points into a separate
`pTCB->xMPUSettings.ulContext[]` array that is used to store the task's
registers for context switches. As a result, computing "Stack Used" and
"Stack Free" as deltas between `pxTopOfStack` and `pxStack` /
`pxEndOfStack` gives incorrect, meaningless results.

* Make stackInfo.stackTop optional, undefine it if MPU context is unusable

On FreeRTOS-MPU for Cortex-M, the location of the stack pointer in the
task MPU context can vary at runtime based on whether a task is using
the FPU or not. The size of the saved context is used to detect whether
that's the case. For any foreground tasks, pxTopOfStack == ulContext so
this check can't be done.

Making stackTop potentially undefined allows Stack Top to be reported as
0x???????? if needed rather than reporting the misleading actual
pxTopOfStack, which doesn't point into the stack at all on ARM MPU
ports.

* Move arch detection routines into rtos-common

In the future, other RTOSes may also need such functionality.