Show vQueueAddToRegistry() hint only for empty queues (#61) (#69)
* Show vQueueAddToRegistry() hint only for empty queues (#61)

The hint is now shown only on the QUEUE and MUTEX tabs, and only if none
of that type of object is registered.

Since it's not necessarily an error for one of those lists to be empty,
soften the language from "missing" to simply "not found in registry".

* Hide QUEUES and MUX/SEMS tabs if configQUEUE_REGISTRY_SIZE == 0

There is a hint at the bottom of the panel explaining that
configQUEUE_REGISTRY_SIZE can be set to > 0 to get information about
queues and semaphores, so the option to see these tabs is still
discoverable. For users who have opted out of the queue registry, this
removes the visual clutter of tabs they can't use.